Make a `q-as-control` layout
Make a `q-as-control` layout which maps levels 5–8 to `Control_L`.  The
purpose of this is the same as for the tab key (†1); to experiment with
a way of making it easier to use the control key in conjunction with
levels 2 and 3.

The reason that I’ve chosen to start this at level 5 is because then I
can access these keys by using the level 5 shift key.  If I used say
level 4, I would need to hold down two keys (level 2 and level 3 shift),
which is less convenient.

†. 1: r#0359f5fbe3e8 (“Make a `tab-as-alt` layout”, 2017-11-04).