Replace `Hyper` keys with group latch keys
I have recently done some research on the `Mode_switch` key, which is an
alias for `ISO_Group_Shift`.  There are group shifting keys that are
analogues to the level 2, 3, and 5 keys: latch, shift, and lock.  In a
way, this expands the repertoire of a keyboard with another key, which
in principle can double the effective levels from eight to sixteen.

I have not felt that ambitious, though.  With my setup I have never had
the need to use a group shift key; I use one layout for programming and
writing English and another for Norwegian.

But it hit me today that there is not only a shift key but also a latch
and a lock key.  Furthermore, `<AE01>` and `<AE11>` keys (or `1` and `+`
on the `no` layout) are in nice positions to act as an additional latch
keys.  These keys are right next to the three other latch keys on either
side.  So with these positions I can combine the group latch keys with
both the other latch keys as well as the compose key.  At this point it
feels a bit opulent… but why not?  I won’t find out if it has potential
unless I try and experiment with it.