Move the function keys to different keys
Move the function keys to the third alphabetic row and the two last to
the `G` and `H` keys, since there are only 10 keys on the third
alphabetic row.

The reason for this move is that having the function keys on the first
alphabetic row does not seem to work well with my keyboard.  I tried to
hold the group latch key and the shift latch key at the same and then
pressing one of the function keys (`GL+SL+F?`) but that didn’t work due
to key jamming.

I don’t know if I will get the same problems with these keys.  I’ll just
have to try and see.