Move `q-as-control` and `tab-as-alt` to group 2
Move `q-as-control` and `tab-as-alt` from levels 5–8 of group 1 to
group 2.

The point of making these was to have access to alternative control and
alt keys on the fifth level, since that can be accessed directly through
modifier 5 (only one modifier key).  This as opposed to the control and
alt keys on the extend layers, which require at least two modifiers to
be held since they are on levels 7 and 8.

Having them on group 2 should make them as easily accessible, since
group 2 is accessed through a single modifier as
well (`ISO_Group_Latch`).

Also remove the now obsolete includes from `no-p6`.