Put `Return` on `M` on the extend levels
Replace `Alt_R` with `Return` on `M` on the extend levels.  Since I
can’t think of a scenario in which I would need `Alt` on this side of
the keyboard on the extend level, and since I am used to typing
`Control+M` in the terminal and in Emacs, I think an `Ext+M` key would
do me more good.