Move parens to third level
Move parens to third level and make other changes to accomodate this.

I don’t want parens to be on the first level since this makes access to
the parens the same on all layouts.  A nice side-effect—and perhaps the
primary motivation?—is that I free up to first level keys on the
English/programming layouts.  I can use these keys for dead keys, for
example.

Also make a few changes that aren’t strictly related, like moving the
percent and dollar sign.