Move the parens again
I think this is a better place for them.  I want them to be convenient
to type.  All I had to give up was the convenience of typing `*` (it’s
now on the `G` key) and I think that was more than worth it, considering
how much more useful it is to be able to type `(` and `)` conveniently
for me.