Add `Hyper` keys
Add `Hyper` keys on the two former Shift Level 3 keys.

I’ve tied these keys to the modifier `NumLock`.  `NumLock` is `Mod2`,
which seemed to be the only modifier that I could spare.  The `Hyper`
keys seemed to be tied to the same modifier as the `Super` keys, namely
`Mod4`.

This interferes with the `NumLock` key.  For example, when `NumLock` is
pressed (the keys on the numpad are available), Emacs will interpret
regular keypresses like `g` (no modifier) as `H-g` (`g` pressed with
`Hyper`).  So with this change I cannot use the `NumLock` key any more.
Which is not a problem in any case since I don’t use the numpad.