Redirect key to keycode
Redirect key to keycode corresponding to `XF86AudioPlay`, the key that
I’m using for playing and pausing (toggling) Spotify.

I play Spotify by using Xchainkeys.  I have this line in
`xchainkeys.conf`:

    XF86AudioPlay :exec playerctl play-pause spotify

So the intent was to map `XF86AudioPlay` to the fifth level of `<UP>`.
This did not work, however.  The reason for that is that Xchainkeys does
not recognize the keysyms of the higher levels.  But by redirecting the
key to another keycode and clearing the modifier, Xchainkeys now
recognizes `mod3+Up` (fifth level of this key) as `XF86AudioPlay`.