Start implementing a “lock unlock” key
Start implementing a key to unlock locked control, alt, and super.  I
want to be able to unlock control, alt, super, and any combination of
the three.  It is simple enough to unlock individual modifiers; just
press the corresponding key.  On the other hand, it does not seem
possible to unlock combinations of locks (e.g. control plus alt) by
pressing only one key or a combination of keys; it seems that I have to
press and release each modifier.

I want to implement a key which can unlock all combinations of control,
alt, and super by pressing and releasing it only once.

§ `ISO_Lock`

This key was added in the previous commit.  This key is used to lock
arbitrary combinations of modifiers.  For example, you can hold down
three modifiers and press `ISO_Lock` in order to lock them:

    ctrl+alt+super+ISO_Lock ⇒ ctrl, alt, and super are locked

What I am implementing now is supposed to complement `ISO_Lock`.