Add type `CTRL_ALT_SUPER`
Add type `CTRL_ALT_SUPER` and in turn the implementation of the “lock
unlocker” key.

The current implementation of the “lock unlocker” key uses key
redirection and the Xbindkeys daemon (†1).  The new implementation will
use Xkb type definitions and no third-party applications (Xbindkeys).

The new type `CTRL_ALT_SUPER` is an eight-level type which uses
`Control`, `Alt`, and `Super` to access levels 2–8.  For each of the
levels 2–8 there will be lock actions for all the involved modifiers.
For example, for keys that use this key type (`CTRL_ALT_SUPER`),
`Level4` is accessed using `Control+Alt`.  The “lock unlocker” key will
have this action on `Level4`:

    action=LockMods(modifiers=Control+Alt);

† 1: See r#f473b5ee48a2 (“Merge branch ‘lock-unlocker’ into next”,
    2019-03-10) for a discussion of the previous implementation.