Change the “lock unlocker”
Assign the `CONTROL_ALT_SUPER` key type and assign all the eight
actions.

Use `VoidSymbol` instead of `Farsi_0`.  This way no symbol gets printed
when I press the `PrintScreen` key without any of the three
modifiers (`Control`, `Alt`, Super`).  The only downside is that Emacs
recognizes `VoidSymbol` and gives the message `<VoidSymbol> is
undefined` in the minibuffer, but I doubt that that will be a problem in
practice.