Change to group 1
Change this layout from group 2 to group 1.  This ought to make it more
flexible, since it should be easier to _use it_ as another group.

If I want to use it as group 2 (the second layout of my setup), simply
pass it to `keyboard-setup` as the second layout:

    /home/kristoffer/xkb/scripts/keyboard-setup \
        "twin-dexter,twin-dexter(all-f-keys)"

If I want to use it as the third group (third layout), simply pass it as
the third argument:

    /home/kristoffer/xkb/scripts/keyboard-setup \
        "twin-dexter,,twin-dexter(all-f-keys)"

And so on.  Much more flexible.

Also remove the inclusion of this layout from `base` since now the group
should be included via `keyboard-setup`.