Try to use `unpack`
I’ve read on StackOverflow that this might be what I am looking for.  So
let’s try it out.

Also make sure that the file that is opened (`keysym-lookup`) is read as
a UTF-8 encoded file.  I don’t really think it matters, though,
considering that there are no actual characters in the file outside of
the ASCII range.  In any case, it doesn’t hurt.