Change the columns in the lookup table
Change the first column to contain Unicode codepoints.

In the process of making the lookup script (`scripts/lookup`), I mixed
things up and tried to interpret the hex numbers in the first column as
UTF-8 codepoints.  I think it will be easier to convert strings like
e.g. `U+2014` to `—` (EM DASH).

Like the first version of this file, this was derived from transforming
the contents of `keysymdef.h`.

§ How this file was derived

This script did the bulk of the work:

    cat reference/keysymdef.h |
        sed -e 's/XK_//' \
        -e 's/#define //' \
        -e 's/ \+/ /g' \
        -e '/U+..../!d' \
        -e '/endif/d' \
        -e '/ifdef/d' \
        -e 's/\(U+....\).*/\1/' \
        -e 's/\/[*](/\/* /' |
        cut --delimiter=" " --fields="1,4" |
        awk -F" " '{print $2,$1}' > keysym-lookup

Then I just had to delete a few lines that were left over (lines from
comments in the original header file).