Replace `☇` with `🔗` for `Q` on level four
I have been using `☇` as a symbol for links and URLs in my own
informal “lightweight markup” language (used mostly for Git commit
messages).  `☇` is more of a meteorological symbol while `🔗` is a very
appropriate symbol for links and URLs (†1).

† 1: “Two links of a silver chain, positioned at a 45° angle. Used as an
    icon for a hyperlink on computers and the internet. May also be used for
    metaphorical connections.” ([Source])

    🔗 Source: https://emojipedia.org/link-symbol/

I have tested this new symbol out in a few applications.  It works in
Firefox, Gnome-terminal and Emacs.  So it seems like it has a wide
enough coverage for my purposes.

§ Background

I might have tried this symbol before.  I recall trying a
link/chain-like symbol before settling on `☇` because `☇` could actually
be rendered in the applications that I cared about.  If it is the case
that I have tried `🔗` before then better support for the symbol must
have been implemented in the meantime.