Map function keys to group 2
I’ve been having a lot of problems with group locking (see the entry in
the journal dated 2019-10-21).  I suspect that it might have something
with how groups are assigned.  Right now I map these functions keys to
group 1 (`all-f-keys`) and pass them in to `keyboard-setup` as the
second group/layout:

    scripts/keyboard-setup "twin-dexter,twin-dexter(all-f-keys)"

So the base layout (`twin-dexter`) is on group 1 and the function
keys (`all-f-keys`) are on group 2.

Maybe this group assignment business is what is causing the group
locking problems?  To test this theory let’s map the function keys to
group 2 directly.  Later I will include this section in a section named
`pr` (for “programming”) and use that as the only layout that I pass to
`keyboard-setup` directly:

    scripts/keyboard-setup "twin-dexter(pr)"